Current Progress
Note: The code has bugs!!

I'm trying to use "facebook/blenderbot-400M-distill" as our foundational model.

When im using PEFT with this model I get an ERROR because the peft wrapping uses the generate function in a way that process input_embeds instead of input_ids. Because the "BlenderbotForConditionalGeneration" that provides this model didnt implement its code this way we get the error.

I'm trying to fix it by extending this class.